requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
beautifulsoup4==4.9.3
black==21.6b0
Brotli==1.0.9
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.8
elasticsearch==7.13.1
fake-useragent==0.1.11
Faker==8.9.0
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.1.0
googletransx==2.4.2
greenlet==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.5.0
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonpointer==2.1
jsonschema==3.2.0
kiwisolver==1.3.1
logparser @ git+https://github.com/my8100/logparser.git@ed7948b271884af68eb3bb13fa9ee51a4892552c
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.3.4
multidict==5.1.0
mypy-extensions==0.4.3
nltk==3.6.7
nmslib==2.0.5
numpy==1.19.5
pandas==1.1.5
parsel==1.6.0
pathspec==0.8.1
pem==21.2.0
pexpect==4.8.0
Pillow==8.3.2
priority==1.3.0
Protego==0.1.16
psutil==5.7.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycares==4.0.0
pycparser==2.20
PyDispatcher==2.0.5
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-slugify==5.0.2
python-socks==1.2.4
python-telegram-bot==13.8.1
pytz==2021.1
queuelib==1.6.1
regex==2021.11.10
requests==2.25.1
rfc3987==1.3.8
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.5.4
Scrapy==2.5.0
scrapy-do==0.5.0
scrapy-fake-useragent==1.4.4
scrapy-splash==0.8.0
scrapyd==1.2.1
scrapyd-client @ git+https://github.com/scrapy/scrapyd-client@4d839039ccd7b84997ba3fd533df313701b5ab68
scrapydweb @ git+https://github.com/my8100/scrapydweb.git@7a3b81dba2cba4279c9465064a693bb277ac20e9
service-identity==21.1.0
six==1.16.0
sklearn==0.0
slack-sdk==3.11.2
soupsieve==2.2.1
spidermon==1.15.0
SQLAlchemy==1.4.20
strict-rfc3339==0.7
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
twint==2.1.20
Twisted==21.2.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
w3lib==1.22.0
webcolors==1.11.1
webscrapingapi==1.2.6
webscrapingapi-scrapy-sdk==1.0.7
Werkzeug==2.0.1
wincertstore==0.2
word2number==1.1
xgboost==1.5.0
XlsxWriter==3.0.2
yarl==1.6.3
zipp==3.4.1
zope.interface==5.4.0
google-cloud-language==2.4.3
transformers==4.19.4
torch==1.11.0